+++ jest-setup.js
@@ -1,5 +1,7 @@
 'use strict';
 
-const getTestRule = require('./getTestRule');
+const getTestRule = require('./getTestRule.js');
+const getTestRuleConfigs = require('./getTestRuleConfigs.js');
 
 global.testRule = getTestRule();
+global.testRuleConfigs = getTestRuleConfigs();